add wrangler script
