+++ frontend/package.json
@@ -8,6 +8,7 @@
     "build:init": "vite build --config vite.script.config.ts",
     "dev": "concurrently \"pnpm run:dev\" \"pnpm watch:script\"",
     "run:dev": "pnpm remix vite:dev",
+    "dev:wrangler": "wrangler pages dev --compatibility-date=2024-02-15 --live-reload --proxy=5100",
     "watch:script": "pnpm run watch:init",
     "watch:init": "vite build --watch --config vite.script.config.ts",
     "start": "pnpm build:init && remix-serve ./build/server/index.js",
@@ -53,7 +54,8 @@
     "concurrently": "^8.2.2",
     "eslint": "^8.57.1",
     "typescript": "5.1.6",
-    "vite-tsconfig-paths": "^4.3.2"
+    "vite-tsconfig-paths": "^4.3.2",
+    "wrangler": "^4.9.1"
   },
   "engines": {
     "node": ">=20.0.0"
